requirements.txt
streamlit
openpyxl
pandas
requests
BeautifulSoup4
google.generativeai
python.dotenv